requirements.txt
# SpendGuard API — Production Dependencies
# Pinned versions for reproducible builds

fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
supabase==2.11.0
python-dotenv==1.0.1
openai==1.65.4
httpx==0.28.1
python-multipart==0.0.20
